+++ src/native/Cargo.toml
@@ -17,6 +17,7 @@ tokio-tungstenite = "0.28.0"
 toml_edit = "0.23.7"
 thiserror = "2.0.17"
 libc = "0.2.177"
+wrapper = { path = "../wrapper" }
 gpu_util = { path = "../gpu_util" }
 tokio = { workspace = true, features = ["macros", "sync"] }
 pyo3 = { workspace = true }
