+++ src/native/src/python/initialize.rs
@@ -23,7 +23,7 @@ pub fn initialize_python(dir: &Dirs) -> Result<Py<PyAny>> {
         let modules = sys.getattr("modules")?;
         let modules = modules.cast::<PyDict>()?;
         let m = PyModule::new(py, "gpu_util")?;
-        gpu_util::gpu_util(&m)?;
+        wrapper::gpu_util(&m)?;
         modules.set_item("gpu_util", m)?;
 
         // プラグインマネージャーのパスをsys.pathに追加
